get rid of another scary looking implicit cast
[git-p4: depot-paths = "//depot/qt/main/": change = 45991]

+++ src/widgets/qbutton.cpp
@@ -504,7 +504,7 @@ void QButton::setAccel( const QKeySequence& key )
 #ifndef QT_NO_ACCEL
     if ( d && d->a )
 	d->a->clear();
-    if ( !key )
+    if ( !(int)key )
 	return;
     ensureData();
     if ( !d->a )
